[Task 635] fix: Fixed build on windows. Signed-off-by: Alexander Kantchev <alexander.kantchev@broadcom.com>
