+++ maven/plugins/actions/src/main/java/com/vmware/pscoe/maven/plugins/JsBasedActionsTestMojo.java
@@ -17,8 +17,9 @@ package com.vmware.pscoe.maven.plugins;
 import java.io.File;
 import java.nio.file.Paths;
 import java.util.List;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
 
-import org.apache.maven.plugins.annotations.*;
 
 @Mojo(name = "run-vro-tests", defaultPhase = LifecyclePhase.TEST)
 public class JsBasedActionsTestMojo extends AbstractVroTestMojo {
@@ -33,8 +34,8 @@ public class JsBasedActionsTestMojo extends AbstractVroTestMojo {
 	protected void addTestbedPaths(List<String> cmd, Configuration config) {
 		String projectRoot = project.getBasedir().toPath().toString();
 		cmd.add("--actions");
-		cmd.add(Paths.get(projectRoot, SRC_JS_PATH).toString());
+		cmd.add(toPathArgument(projectRoot, SRC_JS_PATH));
 		cmd.add("--tests");
-		cmd.add(Paths.get(projectRoot, SRC_TEST_PATH).toString());
+		cmd.add(toPathArgument(projectRoot, SRC_TEST_PATH));
 	}
 }
