+++ maven/plugins/pom.xml
@@ -62,6 +62,11 @@
 			<artifactId>log4j-core</artifactId>
 			<version>2.24.3</version>
 		</dependency>
+		<dependency>
+			<groupId>com.vmware.pscoe.maven.plugins</groupId>
+			<artifactId>common</artifactId>
+			<version>4.2.2-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 
 	<build>
