+++ packages/ecmascript/pom.xml
@@ -14,7 +14,7 @@
   <packaging>package</packaging>
 
   <properties>
-		<main.basedir>${project.baseUri}../../</main.basedir>
+    <main.basedir>${project.baseUri}../../</main.basedir>
     <test.coverage.enabled>true</test.coverage.enabled>
     <test.coverage.reports>text,html,clover,cobertura,lcovonly</test.coverage.reports>
   </properties>
